requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
mediapipe>=0.8.9
tensorflow>=2.8.0
pandas>=1.3.0
scipy>=1.7.0
pillow>=8.0.0

matplotlib~=3.7.5
PyYAML~=6.0.2
setuptools~=68.2.0
torch~=2.4.1
torchvision~=0.19.1